Switch espresso-types dependency from hotshot-query-service to hotshot-query-service-types, allowing light-client to import espresso-types without cyclic dependency
