+++ crates/espresso/types/Cargo.toml
@@ -12,7 +12,7 @@ ignored = ["base64_bytes", "hotshot_testing"]
 doctest = false
 
 [features]
-testing = ["hotshot-query-service/testing"]
+testing = ["hotshot-query-service-types/testing"]
 
 [dependencies]
 alloy = { workspace = true }
@@ -43,7 +43,7 @@ hotshot = { workspace = true }
 hotshot-contract-adapter = { workspace = true }
 hotshot-example-types = { workspace = true }
 hotshot-libp2p-networking = { workspace = true }
-hotshot-query-service = { workspace = true }
+hotshot-query-service-types = { workspace = true }
 hotshot-types = { workspace = true }
 humantime = { workspace = true }
 indexmap = { workspace = true }
@@ -51,7 +51,7 @@ insta = { workspace = true }
 itertools = { workspace = true }
 jf-advz = { workspace = true }
 jf-merkle-tree-compat = { workspace = true }
-jf-utils = { workspace = true }                  # TODO temporary: used only for test_rng()
+jf-utils = { workspace = true }                    # TODO temporary: used only for test_rng()
 lru = { workspace = true }
 moka = { workspace = true }
 num-traits = { workspace = true }
@@ -62,7 +62,7 @@ rand = { workspace = true }
 serde = { workspace = true }
 serde_bytes = { workspace = true }
 serde_json = { workspace = true }
-sha2 = { workspace = true }                      # TODO temporary, used only for VID, should be set in hotshot
+sha2 = { workspace = true }                        # TODO temporary, used only for VID, should be set in hotshot
 sha3 = { workspace = true }
 static_assertions = { workspace = true }
 surf-disco = { workspace = true }
@@ -84,6 +84,7 @@ vid = { workspace = true }
 criterion = "0.5"
 espresso-contract-deployer = { workspace = true }
 espresso-types = { path = ".", features = ["testing"] }
+hotshot-query-service = { workspace = true, features = ["testing"] }
 rand_chacha = { workspace = true }
 rstest = { workspace = true }
 rstest_reuse = { workspace = true }
