+++ crates/espresso/types/src/v0/impls/fee_info.rs
@@ -13,7 +13,7 @@ use espresso_utils::{
     impl_serde_from_string_or_integer, impl_to_fixed_bytes, ser::FromStringOrInteger,
 };
 use hotshot_contract_adapter::sol_types::Deposit;
-use hotshot_query_service::explorer::MonetaryValue;
+use hotshot_query_service_types::explorer::MonetaryValue;
 use hotshot_types::traits::block_contents::BuilderFee;
 use itertools::Itertools;
 use jf_merkle_tree_compat::{
