+++ hotshot-query-service-types/Cargo.toml
@@ -11,6 +11,7 @@ testing = ["hotshot-example-types"]
 
 [dependencies]
 anyhow.workspace = true
+ark-serialize.workspace = true
 committable.workspace = true
 derivative.workspace = true
 derive_more.workspace = true
@@ -19,10 +20,12 @@ hotshot-events-service.workspace = true
 hotshot-types.workspace = true
 itertools.workspace = true
 jf-advz.workspace = true
+jf-merkle-tree-compat.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 snafu.workspace = true
 surf-disco.workspace = true
+tagged-base64.workspace = true
 tide-disco.workspace = true
 time = { workspace = true, features = ["formatting", "parsing"] }
 vbs.workspace = true
